feat(stream): add local status streaming
- Implemented NewStatusStream method in status_stream.go for creating a one-shot local status stream.
- Added comprehensive tests for the status stream in status_stream_test.go, covering handshake, message handling, reconnections, and error scenarios.
- Introduced new error handling structures in errors.go to manage status-stream transport and protocol failures.
- Defined options for configuring the status stream in options.go, including stale data handling and reconnection policies.
- Updated documentation in doc.go to clarify the purpose of the stream package.
- Refactored validation logic in validation.go to improve volume validation.

+++ models.go
@@ -282,11 +282,11 @@ type PlayAudio struct {
 }
 
 type AudioVolumeInfo struct {
-	Volume float64 `json:"volume"`
+	Volume int `json:"volume"`
 }
 
 type SetAudioVolumeRequest struct {
-	Volume float64
+	Volume int
 	Silent bool
 }
 
@@ -308,7 +308,7 @@ type StorageList struct {
 type StorageListElement struct {
 	Type StorageListElementType `json:"type"`
 	Name string                 `json:"name"`
-	Size int64                  `json:"size,omitempty"`
+	Size uint64                 `json:"size,omitempty"`
 }
 
 type StorageListElementType string
@@ -319,9 +319,9 @@ const (
 )
 
 type StorageStatus struct {
-	UsedBytes  int64 `json:"used_bytes"`
-	FreeBytes  int64 `json:"free_bytes"`
-	TotalBytes int64 `json:"total_bytes"`
+	UsedBytes  uint64 `json:"used_bytes"`
+	FreeBytes  uint64 `json:"free_bytes"`
+	TotalBytes uint64 `json:"total_bytes"`
 }
 
 type BusySnapshot struct {
